Major cleanup, some things may still be broken
