+++ img/battlers/decidueye.png
