+++ img/battlers/mamoswine.png
