+++ img/battlers/marowak.png
