+++ img/battlers/quagsire.png
