+++ img/battlers/scizor.png
