+++ img/battlers/smeargle.png
