+++ src/constants.py
@@ -1,0 +1,24 @@
+#Window size
+WINDOW_WIDTH = 1450
+WINDOW_HEIGHT = 800
+
+#Whether or not to use an API (for interfacing with a bot)
+API = False
+
+#Colours
+BACKGROUND = (240, 240, 240)
+
+#Gameplay
+startTimer = 300
+gameOverCountdown = 30
+
+#Sprites
+battler_sprites = [
+    "pikachu","staraptor","infernape","umbreon","mamoswine","nidoking","scizor","wigglytuff",
+    "decidueye","kingdra","smeargle","marowak","quagsire","porygonz","tyranitar","metagross",
+]
+move_sprites = {
+    "bone","bravebird","bubble","dark","dazzling","earthquake","fire1","fire2",
+    "fire3","fire4","fire5","fire6","fire7","fire8", "fist", "ironhead", "irontail", "leaf",
+    "poison", "sandstorm", "shadowball", "stone", "waterfall", "zenheadbutt"
+}
