+++ src/debug.py
@@ -1,0 +1,4 @@
+overrideBattlers = True
+battlerOverride = ["quagsire","nidoking","smeargle"]
+showHitboxes = False
+showCollisionBoxes = False
